requirements.txt
streamlit>=1.28.0
browser-use>=0.1.0
playwright>=1.40.0
playwright-stealth>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
nest-asyncio>=1.5.0
psutil>=5.9.0
pydantic>=2.0.0